requirements.txt
python-dotenv==0.19.2
pymongo==3.13.0
fastapi
uvicorn
pydantic==2.10.2
motor==2.5.1
certifi
aiohttp==3.11.8